ai-service/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
cryptography
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-huggingface
faiss-cpu
python-multipart
sentence-transformers
